Combine script to web scrape cats and one to web scrape items into one script -WIP
